Implement dark mode toggle
Adds a global client-side script in _app.tsx that applies theming
according to the localStorage flag, and falls back to the CSS media
query if it is unset.
It also applies listeners for changes in the CSS media query (due to a
system toggle) or by clicking the sun/moon button added to the navbar.

Adds a 1sec transition for the dark/light colors.