Implement button animations
Redesigns the `background` elements in the policy page buttons to be
used universally in the new <GradientButton> component, which turns that
background into a circle with transitions (animated movement).

GradientButton uses an <a> tag as opposed to a <div> to implement a
:focus state.

Adds button animations to the Policy page and Contact page.
However, no functionality has been defined or implemented for either
page yet.